requirements_upgrade.txt
django-cors-headers
django-debug-toolbar
django-extensions
#django-rest-swagger
djangorestframework
psycopg2
